Create JS patch file
