+++ test/arizona_view_handler_SUITE.erl
@@ -78,6 +78,7 @@ hello_world(Config) when is_list(Config) ->
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Arizona</title>
+        <script src="assets/js/arizona/patch.js"></script>
         <script src="assets/js/arizona/worker.js"></script>
         <script src="assets/js/arizona/main.js"></script>
     </head>
